# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Umalo Video Conference API",
        "description": "API documentation for Umalo video conference application",
        "contact": {
            "email": "support@umalo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get user's rooms",
                "description": "Get all rooms created by the user or public rooms",
                "operationId": "1b492e14cbdc859a446992c209170929",
                "responses": {
                    "200": {
                        "description": "List of rooms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "description": "Create a new meeting room",
                "operationId": "0fcc821623ab708a9cd94310cad7971f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Team Meeting"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Weekly team standup",
                                        "nullable": true
                                    },
                                    "is_public": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "max_participants": {
                                        "type": "integer",
                                        "maximum": 100,
                                        "minimum": 2,
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Room": {
                "title": "Room",
                "description": "Room model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "room_id": {
                        "type": "string",
                        "example": "abc123def456"
                    },
                    "name": {
                        "type": "string",
                        "example": "Team Meeting"
                    },
                    "description": {
                        "type": "string",
                        "example": "Daily standup meeting",
                        "nullable": true
                    },
                    "is_public": {
                        "type": "boolean",
                        "example": false
                    },
                    "max_participants": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_by": {
                        "type": "string",
                        "example": "1"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "ended_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "creator": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}